integrations/django-ninja/requirements.txt
django>=4.2
django-ninja==1.1.0
pydantic==2.7.1
pydantic_core==2.18.2
